==> codes/1_load_News_data.ipynb <==
import os, json
import pandas as pd

##  Function as a generator to load all files in all sub-folders under the parent directory 

def list_files(dirpath):
    for dirname, dirnames, filenames in os.walk(dirpath):
        for filename in filenames:
            yield os.path.join(dirname, filename)

## Load all .json files use json.loads(), basically each .json file only contains 1 line

json_list = []

dirpath = 'C:/Users/YuNan/Downloads/stock/'
for filePath in list_files(dirpath):
     if filePath.endswith('.json'):
            with open(filePath) as f:
                for line in f:
                    data = json.loads(line)
                    for j in range(len([*data])):
                        for k in range(len(data[[*data][j]])):
                            json_list.append([data[[*data][j]][k]['pub_time'], [*data][j], data[[*data][j]][k]['title'], data[[*data][j]][k]['text'], data[[*data][j]][k]['url']])

## Make sure the length of json list matches the total files

len(json_list)

## Convert Json to DataFrame in order to perform data analysis

col_names =  ['published_date','company','title','body','url']
df= pd.DataFrame(json_list,columns=col_names)

df.shape

df.head()

## Sort the data by date

df = df.sort_values(by=['published_date'], ascending=True)

df.head()

df=df.reset_index(inplace=False)
del df['index']

df.head()

## Export to .csv file

df.to_csv('new_data_articles.csv')
