==> codes/3_data_process.ipynb <==
import numpy as np
import pandas as pd

df = pd.read_csv('new_data_articles_with_sentiment.csv',index_col=0)

df.head()

df2=df.copy()
df2=pd.DataFrame(df2)
n=df.shape[0]

for i in range(0,n):
    
    df2['published_date'][i]=df2['published_date'][i][0:10]
    if i%10000==0:
        print(i)
df2.to_csv('sentiment.csv')

In[15]:

df2['company'].unique()

# grouped=df2.groupby(['company']).size()
# print(grouped[:50])

# BA=df2.groupby(['company']).filter(lambda x: x=='BA')
# print(BA)


df2['company'].unique()[0]

for i in range(df2['company'].unique().shape[0]):
    dff1 = df2[df2.company == df2['company'].unique()[i]]
    print(dff1.shape)
    dff1.to_csv('sentiment\sentiment_{}.csv'.format(df2['company'].unique()[i]),index = False)

dff1=df2[(df2.company=='BA')]
print(dff1.shape)
# dff2=df2[(df2.source_name=='cnbc.com')]
# print(dff2.shape)
# dff3=df2[(df2.source_name=='fortune.com')]
# print(dff3.shape)
# dff4=df2[(df2.source_name=='reuters.com')]
# print(dff4.shape)

dff1 = pd.read_csv('sentiment\sentiment_BA.csv')
# dff2.to_csv('D://NeurIPS Workshop/data/sentiment_cnbc.csv')
# dff3.to_csv('D://NeurIPS Workshop/data/sentiment_fortune.csv')
# dff4.to_csv('D://NeurIPS Workshop/data/sentiment_reuters.csv')

dff1.head()

sp = pd.read_csv("C:/Users/YuNan/Downloads/stock/historical_price/historical_price/BA_2015-12-30_2021-02-21_minute.csv")#,index_col=0)
sp=pd.DataFrame(sp)
sp.shape

dff1g=dff1.groupby(['published_date']).agg(['mean'])
print(dff1g.shape)
# dff2g=dff2.groupby(['published_date']).agg(['mean'])
# print(dff2g.shape)
# dff3g=dff3.groupby(['published_date']).agg(['mean'])
# print(dff3g.shape)
# dff4g=dff4.groupby(['published_date']).agg(['mean'])
# print(dff4g.shape)

dff1g.to_csv('sentiment_BA_groupday.csv')
# dff2g.to_csv('D://NeurIPS Workshop/data/sentiment_cnbc_groupday.csv')
# dff3g.to_csv('D://NeurIPS Workshop/data/sentiment_fortune_groupday.csv')
# dff4g.to_csv('D://NeurIPS Workshop/data/sentiment_reuters_groupday.csv')

dff1g.columns

d1 = pd.read_csv('sentiment_BA_groupday.csv')#,index_col=0)
# d2 = pd.read_csv("D://NeurIPS Workshop/data/sentiment_cnbc_groupday1.csv")
# d3 = pd.read_csv("D://NeurIPS Workshop/data/sentiment_fortune_groupday1.csv")
# d4 = pd.read_csv("D://NeurIPS Workshop/data/sentiment_reuters_groupday1.csv")
# d1=pd.DataFrame(d1)
# d1.head()

print(type(sp['t'][0]))
print(type(d1['published_date'][0]))
print(sp.head())
print(d1.head())

sp1=sp.copy()
m=sp.shape[0]

for i in range (0,3):
    print(i)
    sp1['t'][i].replace('-','/')

print(sp1['t'][0])

sp1['t'][0]==d1['published_date'][0]
print(sp1['t'][0])
print(d1['published_date'][0])

import time
t=d1['published_date'][0]
timeStruct = time.strptime(t, "%m/%d/%Y") 
strTime = time.strftime("%Y-%m-%d", timeStruct) 
print(strTime)

for i in range(0,d1.shape[0]):
    t=d1['published_date'][i]
    timeStruct = time.strptime(t, "%m/%d/%Y") 
    d1['published_date'][i] = time.strftime("%Y-%m-%d", timeStruct) 
    
d1.head()
    

# for i in range(0,d2.shape[0]):
#     t=d2['published_date'][i]
#     timeStruct = time.strptime(t, "%Y/%m/%d") 
#     d2['published_date'][i] = time.strftime("%Y-%m-%d", timeStruct) 
    
# for i in range(0,d3.shape[0]):
#     t=d3['published_date'][i]
#     timeStruct = time.strptime(t, "%Y/%m/%d") 
#     d3['published_date'][i] = time.strftime("%Y-%m-%d", timeStruct) 
    
# for i in range(0,d4.shape[0]):
#     t=d4['published_date'][i]
#     timeStruct = time.strptime(t, "%Y/%m/%d") 
#     d4['published_date'][i] = time.strftime("%Y-%m-%d", timeStruct) 

In[17]:

In[30]:

d1.head()

sp1['date'] = sp1['t'].apply(lambda x: x.split(' '))
sp1['Date'] = sp1['date'].apply(lambda x: x[0])
sp1['time'] = sp1['date'].apply(lambda x: x[1])

sp1.drop_duplicates(subset=['Date'], keep='last', inplace = True)
print(sp1.shape)
sp1.head

sp1.reset_index(inplace = True)
sp1

date_union_1=pd.DataFrame(columns=('idx','date','mean_compound','comp_flag'))
sp_len=sp1.shape[0]
d_len=d1.shape[0]
d=d1.copy()
for i in range(0,sp_len):
    idx=i
    date=sp1['Date'][i]
    j=0
    t=0
    while j<d_len:
        if sp1['Date'][i]==d['published_date'][j]:
            mean_compound=d['compound'][j]
            comp_flag=1
            t=1
            break
        j=j+1
    if t==0:
        mean_compound=0
        comp_flag=0
    
    date_union_1=date_union_1.append(pd.DataFrame({'idx':[idx],
                                                  'date':[date],
                                                  'mean_compound':[mean_compound],
                                                  'comp_flag':[comp_flag]
        
    }),ignore_index=True)

date_union_1.shape

date_union_1[600:700]

date_union_1.to_csv('new_date_union_1.csv')
